Upgraded tradermade to v3 framework (#2763)
