+++ .changeset/itchy-cars-tan.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/tradermade-adapter': major
+---
+
+Upgraded adapter from v2 to v3 framework
