--- packages/sources/tradermade-test/CHANGELOG.md
@@ -1,99 +1,0 @@
-# @chainlink/tradermade-test-adapter
-
-## 1.4.0
-
-### Minor Changes
-
-- 6794e4514: Added overrides to the live endpoint
-
-## 1.3.0
-
-### Minor Changes
-
-- ac4dea150: Accounting for credit based rate limiting by unbatching all endpoints
-
-## 1.2.2
-
-### Patch Changes
-
-- bfa201d6d: Bumped framework version
-- bfa201d6d: Bumped framework version
-- c600ca386: Upgrade typescript version to 5.0.4
-- bfa201d6d: Bumped framework version
-
-## 1.2.1
-
-### Patch Changes
-
-- 9bcb13c90: Bumped framework version
-
-## 1.2.0
-
-### Minor Changes
-
-- e0fe0a602: Added custom routing using WS_ENABLED env var
-
-### Patch Changes
-
-- 31af84e69: Bumped framework version
-
-## 1.1.7
-
-### Patch Changes
-
-- 77ad946a9: Bumped framework version
-
-## 1.1.6
-
-### Patch Changes
-
-- 4245b4d8b: Bumped framework version
-
-## 1.1.5
-
-### Patch Changes
-
-- 9eaed42c8: Aligned default endpoint to v2
-- 8fe5be225: Bumped framework version
-
-## 1.1.4
-
-### Patch Changes
-
-- c9cee149e: Bump fw version
-
-## 1.1.3
-
-### Patch Changes
-
-- 2fdaa5aa4: Bump v3 version
-
-## 1.1.2
-
-### Patch Changes
-
-- bc8886c63: Bumped framework version
-
-## 1.1.1
-
-### Patch Changes
-
-- 37c0d8883: Bumped v3 framework version
-
-## 1.1.0
-
-### Minor Changes
-
-- d474a5553: Update v3 framework to latest version
-
-## 1.0.1
-
-### Patch Changes
-
-- 65014014d: Bumped v3 framework version. Upgraded typescript to version 4.9.5
-
-## 1.0.0
-
-### Major Changes
-
-- 49680b472: Initial version of the adapter
