--- packages/sources/tradermade-test/src/config/includes.json
@@ -1,13 +1,0 @@
-[
-  {
-    "from": "IDR",
-    "to": "USD",
-    "includes": [
-      {
-        "from": "USD",
-        "to": "IDR",
-        "inverse": true
-      }
-    ]
-  }
-]
