--- packages/sources/tradermade-test/src/endpoint/index.ts
@@ -1,2 +1,0 @@
-export { endpoint as forex } from './forex-router'
-export { endpoint as live } from './live'
