--- packages/sources/tradermade/schemas/env.json
@@ -1,26 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/tradermade-adapter.json",
-  "title": "Chainlink External Adapter for Tradermade\n\nThis adapter only has Websocket support for the forex endpoint.",
-  "required": ["API_KEY"],
-  "type": "object",
-  "properties": {
-    "API_KEY": {
-      "type": "string",
-      "required": true,
-      "description": "An API key that can be obtained from [here](https://marketdata.tradermade.com/docs/restful-api)"
-    },
-    "API_ENDPOINT": {
-      "default": "https://marketdata.tradermade.com/api/v1/live",
-      "type": "string"
-    },
-    "WS_API_KEY": {
-      "type": "string",
-      "description": "An API key that can be obtained from [here](https://marketdata.tradermade.com/docs/restful-api)"
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
