--- packages/sources/tradermade/src/config/limits.json
@@ -1,23 +1,0 @@
-{
-  "http": {
-    "basic": {
-      "rateLimit1h": 1.369
-    },
-    "professional": {
-      "rateLimit1h": 13.69
-    },
-    "business": {
-      "rateLimit1h": 68.49
-    },
-    "advanced": {
-      "rateLimit1h": 342.46
-    },
-    "enterprise": {
-      "rateLimit1h": 833.33
-    },
-    "enterprise-xl": {
-      "rateLimit1h": 1736.11
-    }
-  },
-  "ws": {}
-}
