--- packages/sources/tradermade/src/config/symbols.json
@@ -1,8 +1,0 @@
-{
-  "tradermade": {
-    "BRENT": "UKOIL",
-    "FTSE": "UK100",
-    "N225": "JPN225",
-    "WTI": "OIL"
-  }
-}
