+++ packages/sources/tradermade/src/endpoint/index.ts
@@ -1,7 +1,2 @@
-import type { TInputParameters as LiveInputParameters } from './live'
-import type { TInputParameters as ForexInputParameters } from './forex'
-
-export type TInputParameters = LiveInputParameters | ForexInputParameters
-
-export * as live from './live'
-export * as forex from './forex'
+export { endpoint as forex } from './forex-router'
+export { endpoint as live } from './live'
