+++ packages/sources/tradermade/test-payload.json
@@ -1,6 +1,6 @@
 {
  "requests": [{
-  "from": "ZAR",
+  "from": "LINK",
   "to": "USD"
  }]
 }
