+++ packages/sources/tradermade/test/integration/adapter.test.ts
@@ -1,213 +1,186 @@
-import { AdapterRequest, FastifyInstance, util } from '@chainlink/ea-bootstrap'
 import request, { SuperTest, Test } from 'supertest'
-import * as process from 'process'
-import { server as startServer } from '../../src'
-import * as nock from 'nock'
-import { AddressInfo } from 'net'
 import {
-  mockForexSingleSuccess,
-  mockLiveSuccess,
-  mockResponseFailure,
-  mockSubscribeResponse,
-  mockUnsubscribeResponse,
-} from './fixtures'
-import {
+  createAdapter,
+  mockForexWebSocketServer,
   mockWebSocketProvider,
-  mockWebSocketServer,
-  MockWsServer,
-  mockWebSocketFlow,
   setEnvVariables,
-} from '@chainlink/ea-test-helpers'
-import { WebSocketClassProvider } from '@chainlink/ea-bootstrap/dist/lib/middleware/ws/recorder'
-import { DEFAULT_WS_API_ENDPOINT } from '../../src/config'
-import { setupExternalAdapterTest } from '@chainlink/ea-test-helpers'
-import type { SuiteContext } from '@chainlink/ea-test-helpers'
+  setupExternalAdapterTest,
+  SuiteContext,
+} from './setup'
+import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
+import { mockForexResponse } from './fixtures'
+import process from 'process'
+import { AddressInfo } from 'net'
+import { Server } from 'mock-socket'
+import { WebSocketClassProvider } from '@chainlink/external-adapter-framework/transports'
+import { sleep } from '@chainlink/external-adapter-framework/util'
 
 describe('execute', () => {
-  const id = '1'
-  const context: SuiteContext = {
-    req: null,
-    server: startServer,
-  }
+  describe('http', () => {
+    let spy: jest.SpyInstance
+    beforeAll(async () => {
+      const mockDate = new Date('2022-01-01T11:11:11.111Z')
+      spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
+    })
 
-  const envVariables = {
-    API_KEY: process.env.API_KEY || 'fake-api-key',
-  }
+    afterAll((done) => {
+      spy.mockRestore()
+      done()
+    })
 
-  setupExternalAdapterTest(envVariables, context)
+    const id = '1'
 
-  describe('forex api', () => {
-    it('should return success for single base/quote pair', async () => {
-      const data: AdapterRequest = {
+    const context: SuiteContext = {
+      req: null,
+      server: async () => {
+        process.env['RATE_LIMIT_CAPACITY_SECOND'] = '100'
+        process.env['METRICS_ENABLED'] = 'false'
+        const server = (await import('../../src')).server
+        return server() as Promise<ServerInstance>
+      },
+    }
+
+    const envVariables = {
+      API_KEY: process.env.API_KEY || 'fake-api-key',
+      WS_API_KEY: process.env.WS_API_KEY || 'fake-api-key',
+    }
+
+    setupExternalAdapterTest(envVariables, context)
+
+    describe('forex endpoint', () => {
+      const data = {
         id,
         data: {
           endpoint: 'forex',
-          base: 'ETH',
-          quote: 'USD',
+          base: 'eth',
+          quote: 'usd',
         },
       }
 
-      mockForexSingleSuccess()
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
-    })
+      it('should return success', async () => {
+        mockForexResponse()
 
-    // NOTE: batching currently disabled, awaiting service agreement
-    //   it('should return success for batched base/quote pairs', async () => {
-    //     const data: AdapterRequest = {
-    //       id,
-    //       data: {
-    //         endpoint: 'forex',
-    //         base: ['ETH', 'BTC'],
-    //         quote: ['USD', 'JPY'],
-    //       },
-    //     }
-
-    //     mockForexBatchedSuccess()
-    //     const response = await req
-    //       .post('/')
-    //       .send(data)
-    //       .set('Accept', '*/*')
-    //       .set('Content-Type', 'application/json')
-    //       .expect('Content-Type', /json/)
-    //       .expect(200)
-    //     expect(response.body).toMatchSnapshot()
-    //   })
-  })
-
-  describe('live  api', () => {
-    const data: AdapterRequest = {
-      id,
-      data: {
-        endpoint: 'live',
-        base: 'AAPL',
-      },
-    }
-
-    it('should return success', async () => {
-      mockLiveSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
-    })
-  })
-
-  describe('live  api with invalid base', () => {
-    const data: AdapterRequest = {
-      id,
-      data: {
-        endpoint: 'live',
-        base: 'NON-EXISTING',
-      },
-    }
-
-    it('should return failure', async () => {
-      mockResponseFailure()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(502)
-      expect(response.body).toMatchSnapshot()
-    })
-  })
-})
-
-describe('websocket', () => {
-  let mockedWsServer: InstanceType<typeof MockWsServer>
-  let fastify: FastifyInstance
-  let req: SuperTest<Test>
-
-  let oldEnv: NodeJS.ProcessEnv
-  beforeAll(async () => {
-    if (!process.env.RECORD) {
-      process.env.API_KEY = 'fake-api-key'
-      process.env.WS_API_KEY = 'fake-api-key'
-      mockedWsServer = mockWebSocketServer(DEFAULT_WS_API_ENDPOINT)
-      mockWebSocketProvider(WebSocketClassProvider)
-    }
-
-    oldEnv = JSON.parse(JSON.stringify(process.env))
-    process.env.WS_ENABLED = 'true'
-    process.env.WS_SUBSCRIPTION_TTL = '100'
-
-    fastify = await startServer()
-    req = request(`localhost:${(fastify.server.address() as AddressInfo).port}`)
-  })
-
-  afterAll((done) => {
-    setEnvVariables(oldEnv)
-    nock.restore()
-    nock.cleanAll()
-    nock.enableNetConnect()
-    fastify.close(done)
-  })
-
-  describe('forex endpoint', () => {
-    const jobID = '1'
-
-    it('should return success', async () => {
-      const data: AdapterRequest = {
-        id: jobID,
-        data: {
-          endpoint: 'forex',
-          base: 'ETH',
-          quote: 'USD',
-        },
-      }
-
-      let flowFulfilled = Promise.resolve(true)
-      if (!process.env.RECORD) {
-        mockForexSingleSuccess() // For the first response
-
-        flowFulfilled = mockWebSocketFlow(mockedWsServer, [
-          mockSubscribeResponse,
-          mockUnsubscribeResponse,
-        ])
-      }
-
-      const makeRequest = () =>
-        req
+        const response = await (context.req as SuperTest<Test>)
           .post('/')
           .send(data)
           .set('Accept', '*/*')
           .set('Content-Type', 'application/json')
           .expect('Content-Type', /json/)
           .expect(200)
+        expect(response.body).toMatchSnapshot()
+      })
+    })
 
-      // We don't care about the first response, coming from http request
-      // This first request will start both batch warmer & websocket
-      await makeRequest()
+    describe('live endpoint', () => {
+      it('should return success', async () => {
+        mockForexResponse()
 
-      // This final request should disable the cache warmer, sleep is used to make sure that the data is  pulled from the websocket
-      // populated cache entries.
-      await util.sleep(100)
-      const response = await makeRequest()
+        const data = {
+          id,
+          data: {
+            endpoint: 'live',
+            base: 'aapl',
+          },
+        }
 
-      expect(response.body).toEqual({
-        jobRunID: '1',
-        result: 2797.835,
-        statusCode: 200,
-        maxAge: 30000,
-        data: { result: 2797.835 },
+        const response = await (context.req as SuperTest<Test>)
+          .post('/')
+          .send(data)
+          .set('Accept', '*/*')
+          .set('Content-Type', 'application/json')
+          .expect('Content-Type', /json/)
+          .expect(200)
+        expect(response.body).toMatchSnapshot()
       })
 
-      await flowFulfilled
-    }, 30000)
+      it('override should return success', async () => {
+        mockForexResponse()
+
+        const data = {
+          id,
+          data: {
+            endpoint: 'live',
+            base: 'WTI',
+            quote: 'USD',
+          },
+        }
+
+        const response = await (context.req as SuperTest<Test>)
+          .post('/')
+          .send(data)
+          .set('Accept', '*/*')
+          .set('Content-Type', 'application/json')
+          .expect('Content-Type', /json/)
+          .expect(200)
+        expect(response.body).toMatchSnapshot()
+      })
+    })
+  })
+
+  describe('websocket', () => {
+    describe('forex endpoint', () => {
+      let fastify: ServerInstance | undefined
+      let req: SuperTest<Test>
+      let mockWsServer: Server | undefined
+      let spy: jest.SpyInstance
+      const wsEndpoint = 'ws://localhost:9090'
+
+      jest.setTimeout(10000)
+
+      const data = {
+        data: {
+          base: 'ETH',
+          quote: 'USD',
+          endpoint: 'forex',
+          transport: 'ws',
+        },
+      }
+
+      let oldEnv: NodeJS.ProcessEnv
+      beforeAll(async () => {
+        oldEnv = JSON.parse(JSON.stringify(process.env))
+        process.env['WS_SUBSCRIPTION_TTL'] = '10000'
+        process.env['CACHE_MAX_AGE'] = '10000'
+        process.env['CACHE_POLLING_MAX_RETRIES'] = '0'
+        process.env['METRICS_ENABLED'] = 'false'
+        process.env['WS_API_KEY'] = 'fake-api-key'
+        process.env['API_KEY'] = 'fake-api-key'
+        process.env['WS_API_ENDPOINT'] = wsEndpoint
+        process.env['WS_ENABLED'] = 'true'
+        const mockDate = new Date('2022-11-11T11:11:11.111Z')
+        spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
+
+        mockWebSocketProvider(WebSocketClassProvider)
+        mockWsServer = mockForexWebSocketServer(wsEndpoint)
+
+        fastify = await expose(createAdapter())
+        req = request(`http://localhost:${(fastify?.server.address() as AddressInfo).port}`)
+
+        // Send initial request to start background execute
+        await req.post('/').send(data)
+        await sleep(5000)
+      })
+
+      afterAll((done) => {
+        spy.mockRestore()
+        setEnvVariables(oldEnv)
+        mockWsServer?.close()
+        fastify?.close(done())
+      })
+
+      it('should return success', async () => {
+        const makeRequest = () =>
+          req
+            .post('/')
+            .send(data)
+            .set('Accept', '*/*')
+            .set('Content-Type', 'application/json')
+            .expect('Content-Type', /json/)
+            .expect(200)
+
+        const response = await makeRequest()
+        expect(response.body).toMatchSnapshot()
+      }, 30000)
+    })
   })
 })
