+++ packages/tsconfig.json
@@ -540,9 +540,6 @@
       "path": "./sources/tradermade"
     },
     {
-      "path": "./sources/tradermade-test"
-    },
-    {
       "path": "./sources/tradingeconomics"
     },
     {
